+++ packages/extensions/lib/src/http/logging/http_client_async_logger.dart
@@ -1,5 +1,3 @@
-import 'dart:io' show HttpClient;
-
 import 'package:http/http.dart';
 
 import '../../../http.dart' show HttpClientFactory;
@@ -7,8 +5,8 @@ import '../../../system.dart';
 import '../http_client_factory.dart' show HttpClientFactory;
 import 'http_client_logger.dart';
 
-/// An abstraction for asyncronous custom HTTP request logging for a
-/// named [HttpClient] instances returned by [HttpClientFactory].
+/// An abstraction for asyncronous custom HTTP request logging for named
+/// HTTP client instances returned by [HttpClientFactory].
 abstract class HttpClientAsyncLogger extends HttpClientLogger {
   Future<Object?> logRequestStartAsync(
     BaseRequest request,