+++ packages/extensions/lib/src/http/logging/http_client_logger.dart
@@ -1,12 +1,10 @@
-import 'dart:io' show HttpClient;
-
 import 'package:http/http.dart';
 
 import '../../../http.dart' show HttpClientFactory;
 
 import '../http_client_factory.dart' show HttpClientFactory;
 
-/// An abstraction for custom HTTP request logging for a named [HttpClient]
+/// An abstraction for custom HTTP request logging for named HTTP client
 /// instances returned by [HttpClientFactory].
 abstract class HttpClientLogger {
   /// Logs before sending an HTTP request.